requirements.txt
streamlit>=1.28.0
openai>=1.3.0
Pillow>=10.0.0
python-dotenv>=1.1.1